rename: 플레이어 관련 객체 패키지 이동
