+++ src/main/java/domain/player/Name.java
@@ -1,8 +1,6 @@
-package domain;
+package domain.player;
 
-import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
 
 public class Name {
