+++ src/test/java/domain/NameTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.player.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
