+++ src/test/java/domain/NamesTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.player.Names;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
