+++ src/test/java/domain/PlayerTest.java
@@ -3,6 +3,9 @@ package domain;
 import static org.assertj.core.api.Assertions.*;
 
 import domain.ladder.Step;
+import domain.player.Name;
+import domain.player.Player;
+import domain.player.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
