+++ src/test/java/domain/PlayersTest.java
@@ -2,6 +2,8 @@ package domain;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import domain.player.Names;
+import domain.player.Players;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
