+++ src/test/java/domain/StepTest.java
@@ -1,6 +1,7 @@
 package domain;
 
 import domain.ladder.Step;
+import domain.player.Position;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
